requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aioredis==2.0.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
atari-py==0.2.9
atomicwrites==1.4.0
attrs==22.2.0
beautifulsoup4==4.11.2
blessings==1.7
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
click==8.1.2
cloudpickle==2.2.1
colorama==0.4.5
colorful==0.5.4
configparser==5.0.1
contextvars==2.4
cycler==0.11.0
Cython==0.29.33
deepdiff==6.3.0
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
fasteners==0.18
filelock==3.3.2
funcsigs==1.0.2
future==0.16.0
gast==0.2.2
gin==0.1.6
gin-config==0.3.0
gitdb==4.0.10
GitPython==3.1.31
glfw==1.12.0
google==3.0.0
google-api-core==1.31.1
google-auth
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gpustat==0.6.0
gql==3.4.1
graphql-core==3.2.3
grpcio==1.51.3
gym==0.26.2
h5py==3.7.0
hiredis==2.2.3
idna==3.3
idna-ssl==1.1.0
imageio==2.27.0
immutables==0.19
importlib-metadata==5.2.0
joblib==1.2.0
jsonnet==0.19.1
jsonpickle==3.0.1
jsonschema==4.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lockfile==0.12.2
Markdown==3.4.3
matplotlib==3.7.0
mkl-fft==1.3.6
mkl-random==1.2.2
mock==5.0.1
monotonic==1.6
more-itertools==9.1.0
mpi4py==3.1.3
mpyq==0.2.5
msgpack==1.0.4
multidict==6.0.4
munch==2.5.0
numpy==1.24.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencensus==0.7.13
opencensus-context==0.1.2
opt-einsum==3.2.1
ordered-set==4.1.0
packaging==23.1
pandas==2.0.0
pathlib2==2.3.7
pathtools==0.1.2
pbr==5.11.0
Pillow==9.5.0
pluggy==0.13.1
portpicker==1.3.1
probscale==0.2.5
progressbar2==3.53.1
prometheus-client==0.16.0
promise==2.3
protobuf==4.21.3
psutil==5.9.4
py==1.11.0
py-spy==0.3.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.1.2
pyglet==1.5.21
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.3
PySC2==3.0.0
pytest==7.0.1
python-dateutil==2.8.2
python-utils==3.6.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s2clientprotocol==4.10.1.75800.0
s2protocol==4.11.4.78285.0
sacred==0.8.2
scipy==1.10.1
seaborn==0.12.1
sentry-sdk==1.18.0
setproctitle==1.3.1
shortuuid==1.0.1
six==1.16.0
sk-video==1.1.10
smmap==5.0.0
snakeviz==1.0.0
soupsieve==2.4.1
subprocess32==3.5.4
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.2.0
tornado==6.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
wandb==0.15.4
watchdog==2.1.2
websocket-client==1.6.1
Werkzeug==2.3.6
whichcraft==0.6.1
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.8.2
zipp==3.14.0
zmq==0.0.0
pytorch-ignite==0.5.0.post2
omegaconf==2.0.6
scikit_learn==1.3.0